server/requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-environ==0.4.5
django-extensions==3.1.1
flake8==3.9.0
importlib-metadata==3.7.3
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.1
